+++ lib/roo_on_rails/checks/documentation/playbook.rb
@@ -1,4 +1,5 @@
 require 'roo_on_rails/checks/base'
+require 'fileutils'
 
 module RooOnRails
   module Checks
@@ -11,11 +12,28 @@ module RooOnRails
         end
 
         def call
-          if File.exist?(LOCATION)
-            pass 'playbook found, legion on-call engineers thank you.'
-          else
-            fail! "please create a playbook at #{LOCATION}."
-          end
+          fail! "no playbook at #{LOCATION}." if playbook_missing?
+          final_fail! 'playbook still contains FIXME template sections' if playbook_unfinished?
+          pass 'playbook found, legion on-call engineers thank you.'
+        end
+
+        def fix
+          FileUtils.cp(
+            File.join(__dir__, 'playbook_template.md'),
+            LOCATION
+          )
+        end
+
+        private
+
+        def playbook_unfinished?
+          # The regexp is so that you can still refer to strings saying FIXME in your readme
+          # if you need to, by putting the phrase in backticks: `FIXME`
+          !File.read(LOCATION).match(/FIXME(?!`)/).nil?
+        end
+
+        def playbook_missing?
+          !File.exist?(LOCATION)
         end
       end
     end