+++ lib/roo_on_rails/config.rb
@@ -17,6 +17,22 @@ module RooOnRails
         enabled? 'GOOGLE_AUTH_ENABLED', default: false
       end
 
+      def google_auth_path_prefix
+        ENV.fetch('GOOGLE_AUTH_PATH_PREFIX')
+      end
+
+      def google_auth_controller
+        ENV.fetch('GOOGLE_AUTH_CONTROLLER')
+      end
+
+      def routemaster_enabled?
+        enabled? 'ROUTEMASTER_ENABLED', default: false
+      end
+
+      def routemaster_publishing_enabled?
+        enabled? 'ROUTEMASTER_PUBLISHING_ENABLED', default: false
+      end
+
       private
 
       ENABLED_PATTERN = /\A(YES|TRUE|ON|1)\Z/i